backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
openai
pydantic-settings 